@@ -3,6 +3,8 @@ package com.elkdev.tripPlanner.user.controller;
 
 import com.elkdev.tripPlanner.user.model.User;
 import com.elkdev.tripPlanner.user.service.UserService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -10,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Optional;
 
+@Api(description = "Users resource")
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
 public class UserController {
@@ -23,7 +26,7 @@ public class UserController {
         this.bcryptEncoder = bcryptEncoder;
     }
 
-
+    @ApiOperation(value = "Get users")
     @RequestMapping(value="/users", method = RequestMethod.GET)
     public List<User> listUser(){
         return userService.findAll();
@@ -34,6 +37,7 @@ public class UserController {
         return userService.findById(id);
     }
 
+    @ApiOperation(value = "Create new user")
     @RequestMapping(value="/signup", method = RequestMethod.POST)
     public User saveUser(@RequestBody User user){
         user.setPassword(bcryptEncoder.encode(user.getPassword()));
