@@ -1,21 +1,23 @@
 package com.elkdev.tripPlanner.user.model;
 
+import com.elkdev.tripPlanner.country.model.Country;
 import com.elkdev.tripPlanner.security.model.UserRole;
+import com.elkdev.tripPlanner.trip.model.Trip;
 import lombok.*;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
+import java.util.List;
+import java.util.Set;
 
 @Entity(name = "users")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@EqualsAndHashCode
 public class User {
 
     @Id
-    @GeneratedValue
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
     private String lastname;
@@ -24,5 +26,9 @@ public class User {
     private String password;
     private Boolean enabled;
     private UserRole userRole;
+    @ManyToMany
+    private List<Country> userCountries;
+    @OneToMany
+    private List<Trip> userTrips;
 
 }
